Rename newtondynamics to newtondynamics3
